+++ src/joseki/__init__.py
@@ -1,13 +1,15 @@
 """Joseki."""
 from . import accessor
 from .__version__ import __version__ as __version__
-from .core import make
+from .core import make, open_dataset, load_dataset
 from .units import ureg as unit_registry
 
 
 __all__ = [
     "accessor",
+    "load_dataset",
     "make",
+    "open_dataset",
     "unit_registry",
     "__version__",
 ]